rename PageContent to Index
